fix: better error loading of char data
The user-defined function is now given a reject callback to trigger error
conditions while loading new character data. Please note that the
running code (not the user-defined function) is also given a chance to
reject the promise (and thus trigger the calling of the error callback)
when it receives a second loading request before the current one is
finished.